[0.0.2] don’t create package if exists; don’t deploy if already deployed

+++ index.js
@@ -21,30 +21,35 @@ task('package', {async: true}, function () {
   ]).then((commit_and_basedir) => {
     var commit = commit_and_basedir[0];
     var basedir = commit_and_basedir[1];
+    var packageFileName = `package_${Path.basename(basedir)}_${commit}.zip`;
+
     return new Promise((resolve, reject) => {
-      var packageFileName = `package_${Path.basename(basedir)}_${commit}.zip`;
-      process.chdir(Tempy.directory());
-      jake.exec([
-        `rsync -a ${basedir}/.git .`,
-        `git reset --hard ${commit}`,
-        "npm install --production",
-        "zip -r package.zip index.js node_modules",
-        `cp package.zip ${basedir}/${packageFileName}`,
-      ], {printStdout: true, printStderr: true}, () => {
-        resolve({commit, packageFileName});
-      });
-    }).then((result) => {
-      console.log(`Package created in ${result.packageFileName}`);
-      complete({commit: result.commit, path: `${basedir}/${result.packageFileName}`});
+      const fs = require('fs');
+
+      if (fs.existsSync(packageFileName)) {
+        resolve();
+      } else {
+        process.chdir(Tempy.directory());
+        jake.exec([
+          `rsync -a ${basedir}/.git .`,
+          `git reset --hard ${commit}`,
+          "npm install --production",
+          "zip -r package.zip index.js node_modules",
+          `cp package.zip ${basedir}/${packageFileName}`,
+        ], {printStdout: true, printStderr: true}, () => {
+          console.log(`Package created in ${packageFileName}`);
+          resolve();
+        });
+      }
+    }).then(() => {
+      complete({commit: commit, path: `${basedir}/${packageFileName}`});
     });
   });
 });
 
 desc("Deploys the package on AWS.");
-task('deploy', ['package'], {async: true}, function(functionName) {
-  if (!functionName) {
-    throw "You have to specify a function name to deploy to.";
-  }
+task('deploy', ['package'], {async: true}, function(FunctionName) {
+  if (!FunctionName) { fail("You have to specify a function name to deploy to."); }
 
   const Aws = require('aws-sdk');
   const fs = require('fs');
@@ -52,17 +57,26 @@ task('deploy', ['package'], {async: true}, function(functionName) {
 
   var package = jake.Task['package'].value;
   var lambda = new Aws.Lambda({region: 'eu-west-1'});
-  lambda.updateFunctionCode({
-    FunctionName: functionName,
-    ZipFile: fs.readFileSync(package.path),
-  }).promise().then(() => {
-    return lambda.publishVersion({
-      FunctionName: functionName,
-      Description: package.commit,
+  lambda.listAliases({FunctionName}).promise().then((result) => {
+    return result.Aliases.find((alias) => { return alias.Name == 'active'; });
+  }).then((activeAlias) => {
+    return lambda.listVersionsByFunction({FunctionName}).promise().then((result) => {
+      var activeVersion = result.Versions.find(
+          (version) => { return version.Version == activeAlias.FunctionVersion; });
+      if (activeVersion.Description == package.commit) {
+        throw `Commit ${package.commit} is already deployed.`;
+      }
+    });
+  }).then(() => {
+    return lambda.updateFunctionCode({
+      FunctionName,
+      ZipFile: fs.readFileSync(package.path)
     }).promise();
+  }).then(() => {
+    return lambda.publishVersion({FunctionName, Description: package.commit}).promise();
   }).then((result) => {
     return lambda.updateAlias({
-      FunctionName: functionName,
+      FunctionName,
       Name: 'active',
       FunctionVersion: result.Version,
     }).promise();
@@ -92,7 +106,7 @@ task('deploy', ['package'], {async: true}, function(functionName) {
           resolve(stdout.toString().trim());
         });
       }).then((commitMsg) => {
-        var announcement = `:aws: :lambda: (${functionName}) ` +
+        var announcement = `:aws: :lambda: (${FunctionName}) ` +
             `@${process.env.SLACK_USER || process.env.USER} deployed ` +
             `“<${commitUrl}|${commitMsg.replace(">", "&gt;")}>”.`;
         return slackr.string(announcement).then(() => {
@@ -103,5 +117,7 @@ task('deploy', ['package'], {async: true}, function(functionName) {
       console.log(`Could not notify DRI channel: ${error}`);
       console.log("Don't forget to post an announcement there.");
     });
-  });
+  }).catch((e) => {
+    console.log(`Didn't deploy: ${e}`);
+  }).then(() => { complete(); });
 });
